refactor: remove superfluous btop prefix from source files
Closes: https://github.com/aristocratos/btop/issues/1348
