+++ src/config.cpp
@@ -31,10 +31,10 @@ tab-size = 4
 #include <fmt/core.h>
 #include <sys/statvfs.h>
 
-#include "btop_config.hpp"
-#include "btop_log.hpp"
-#include "btop_shared.hpp"
-#include "btop_tools.hpp"
+#include "config.hpp"
+#include "log.hpp"
+#include "shared.hpp"
+#include "tools.hpp"
 
 using std::array;
 using std::atomic;
