+++ src/shared.hpp
