+++ src/tools.hpp
@@ -52,7 +52,7 @@ tab-size = 4
 #include <fmt/core.h>
 #include <fmt/format.h>
 
-#include "btop_log.hpp"
+#include "log.hpp"
 
 using std::array;
 using std::atomic;
